requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.2
mutagen==1.46.0
pycparser==2.21
pycryptodomex==3.17
PyNaCl==1.5.0
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
websockets==10.4
yarl==1.8.1
yt-dlp
